requirements.txt
streamlit>=1.12.0
google-cloud-dialogflow==2.41.1